Make time comparisons more robust in homa_peer.c
Change peertab->idle_jiffies_min and peertab->idle_jiffies_max
from unsigned long to int. This prevents undesirable type
promtions when comparing to signed clock differences. In practice
this should have no impact, since the values being compared
against (in homa_peer_pick_victims) should never be negative,
but this is the arithmetically "correct" thing to do.

+++ homa_peer.h
@@ -73,7 +73,7 @@ struct homa_peertab {
 	 * @idle_jiffies_min: Same as idle_secs_min except in units
 	 * of jiffies.
 	 */
-	unsigned long idle_jiffies_min;
+	int idle_jiffies_min;
 
 	/**
 	 * @idle_secs_max: A peer that has been idle for less than
@@ -87,7 +87,7 @@ struct homa_peertab {
 	 * @idle_jiffies_max: Same as idle_secs_max except in units
 	 * of jiffies.
 	 */
-	unsigned long idle_jiffies_max;
+	int idle_jiffies_max;
 
 #ifndef __STRIP__ /* See strip.py */
 	/**
